backend/requirements.txt
# FastAPI and web server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# LangChain core
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-openai==0.2.5
langchain-text-splitters==0.3.2
langchain-huggingface==1.2.0

# LangGraph
langgraph>=0.2.45
langgraph-checkpoint>=2.0.2
langgraph-checkpoint-postgres>=2.0.6

# Vector store and embeddings
langchain-pinecone==0.2.0
pinecone-client==5.0.1
sentence-transformers==3.3.1

# Document processing
pypdf==5.1.0

# Search
tavily-python==0.5.0

# Database
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Utilities
python-dotenv==1.0.1